% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_cols.R
\name{na_cols}
\alias{na_cols}
\title{Returns cols with NA as string or vector}
\usage{
na_cols(df, show = FALSE)
}
\arguments{
\item{df}{A dataframe.}

\item{show}{A boolean.}
}
\value{
If \code(show), returns a string. Else, it returns a vector
}
\description{
Returns cols with NA as string or vector
}
\examples{
na_cols(df)
na_cols(df, show = TRUE)


}
