% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rows.R
\name{na_rows}
\alias{na_rows}
\title{Returns rows with NA as string or dataframe}
\usage{
na_rows(df, show = FALSE)
}
\arguments{
\item{df}{A dataframe.}

\item{show}{A boolean.}
}
\value{
If \code(show), returns a string. Else, it returns a dataframe
}
\description{
Returns rows with NA as string or dataframe
}
\examples{
na_rows(df)
na_rows(df, show = TRUE)


}
